+++ public/styles.css
@@ -4,4 +4,5 @@ body {
   width: 100vw;
   height: 100vh;
   font-family: Pretendard;
+  overflow-y: overlay;
 }
